+++ internal/generated_file_test/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
+
+package(default_visibility = ["//visibility:public"])
+
+bzl_library(
+    name = "bzl",
+    srcs = glob(["*.bzl"]),
+    visibility = ["//visibility:public"],
+)
+
+exports_files(["bin.js"])
+
+filegroup(
+    name = "package_contents",
+    srcs = glob(["*"]),
+)
