--- internal/golden_file_test/BUILD.bazel
@@ -1,8 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-exports_files(["bin.js"])
-
-filegroup(
-    name = "package_contents",
-    srcs = glob(["*"]),
-)
