+++ internal/npm_install/test/BUILD.bazel
@@ -20,7 +20,7 @@ jasmine_node_test(
 )
 
 nodejs_binary(
-    name = "test.accept",
+    name = "test.update",
     data = [
         ":check.js",
         ":goldens",
