+++ packages/terser/test/debug/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@build_bazel_rules_nodejs//internal/golden_file_test:golden_file_test.bzl", "golden_file_test")
+load("@build_bazel_rules_nodejs//:index.bzl", "generated_file_test")
 load("//packages/terser:index.bzl", "terser_minified")
 
 terser_minified(
@@ -9,10 +9,10 @@ terser_minified(
     sourcemap = False,
 )
 
-golden_file_test(
+generated_file_test(
     name = "test",
-    actual = "out.min",
-    golden = "output.debug.golden.js_",
+    src = "output.debug.golden.js_",
+    generated = "out.min",
 )
 
 terser_minified(
@@ -23,9 +23,9 @@ terser_minified(
     # Instead we'll run the test with --compilation_mode=dbg
 )
 
-golden_file_test(
+generated_file_test(
     name = "test_define_DEBUG",
-    actual = "debug_from_env",
-    golden = "output.golden.js_",
-    golden_debug = "output.debug.golden.js_",
+    src = "output.golden.js_",
+    generated = "debug_from_env",
+    src_dbg = "output.debug.golden.js_",
 )
