+++ packages/typescript/test/ts_project/simple/BUILD.bazel
@@ -1,13 +1,13 @@
-load("@build_bazel_rules_nodejs//internal/golden_file_test:golden_file_test.bzl", "golden_file_test")
+load("@build_bazel_rules_nodejs//:index.bzl", "generated_file_test")
 load("//packages/typescript:index.bzl", "ts_project")
 
 # This uses defaults for all attributes.
 # It will find `index.ts` and produce `index.js`
 ts_project()
 
-golden_file_test(
+generated_file_test(
     name = "test",
+    src = "index.golden.js",
     # Refers to the output from ts_project above
-    actual = "index.js",
-    golden = "index.golden.js",
+    generated = "index.js",
 )
